+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - You can now configure this plugin to run as a platform with multiple TVs
+- Small code optimizations
+- Additional checks to make sure required properties are set
 
 ### Changed